backend/password-reset/requirements.txt
psycopg2-binary==2.9.9
bcrypt==4.1.2
requests==2.31.0